add empty state for android widget
