+++ android/app/src/main/res/layout/courses_widget_empty.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    style="@style/Widget.Android.AppWidget.Container"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:theme="@style/Theme.Android.AppWidgetContainer">
+
+    <TextView
+        android:id="@+id/empty_text"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="You have no classes today 😎"
+        android:textSize="16sp"
+        android:textStyle="bold"
+        android:layout_centerInParent="true">
+    </TextView>
+
+</RelativeLayout>
